src/main/java/mc/owls/valley/net/feathercore/modules/reload/components/ReloadModule.java
/**
 * ------------------------------------------------------------------------- *
 *                     Copyright (c) by FeatherCore 2024                     *
 * ------------------------------------------------------------------------- *
 * @license https://github.com/TheAncientOwl/feather-core/blob/main/LICENSE
 *
 * @file ReloadModule.java
 * @author Alexandru Delegeanu
 * @version 0.4
 * @description Dummy module for registering @see ReloadCommand.java
 */

package mc.owls.valley.net.feathercore.modules.reload.components;

import mc.owls.valley.net.feathercore.api.core.FeatherModule;
import mc.owls.valley.net.feathercore.api.exceptions.FeatherSetupException;
import mc.owls.valley.net.feathercore.modules.reload.interfaces.IReloadModule;

public class ReloadModule extends FeatherModule implements IReloadModule {

    public ReloadModule(final InitData data) {
        super(data);
    }

    @Override
    protected void onModuleEnable() throws FeatherSetupException {}

    @Override
    protected void onModuleDisable() {}

}
